refactor: instance 파일에서 api, token 관련함수 분리
